Registry for simple study importers
